+++ docs/source/release_notes.rst
@@ -8,6 +8,7 @@ Release Notes
     * Changes
     * Documentation Changes
         * Fixed and updated code blocks in Release Notes :pr:`1243`
+        * Added DecisionTree estimators to API Reference :pr:`1246`
     * Testing Changes