+++ docs/source/release_notes.rst
@@ -33,7 +33,7 @@ Release Notes
         * Added install documentation for `libomp` in order to use LightGBM on Mac :pr:`1233`
         * Improved description of `max_iterations` in documentation :pr:`1212`
         * Removed unused code from sphinx conf :pr:`1235`
-        * Changed class inheritance display to flow vertically :pr:``
+        * Changed class inheritance display to flow vertically :pr:`1248`
     * Testing Changes
 
 .. warning::